Highly Complete SNSF: Fix lorom mirroring crash
This bug would only have appeared on 64 bit architectures, due to a
weird casting issue with the return of CMemory::map_mirror and how the
code decides to perform pointer arithmetic. With the incorrect casting
behavior, it ends up wrapping the offset in the 32 bit space and
indexing way past the end of the ROM buffer.

I'm surprised nobody caught this at any other time or port of Snes9x.

+++ Frameworks/snes9x/snes9x/memmap.cpp
@@ -678,7 +678,7 @@ void CMemory::InitROM()
 
 // memory map
 
-uint32_t CMemory::map_mirror(uint32_t size, uint32_t pos)
+ssize_t CMemory::map_mirror(uint32_t size, uint32_t pos)
 {
 	// from bsnes
 	if (!size)
